♻️ refactor: 루시드 아이콘으로 아이콘 이미지 변경
