+++ src/app/shopping/search/page.tsx
@@ -1,7 +1,7 @@
 import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
 import AllItems from '@/app/shopping/AllItems';
-import Search from '@/components/common/Search';
+import SearchBar from '@/components/common/SearchBar';
 
 export default function ShoppingSearch() {
   return (
@@ -20,7 +20,7 @@ export default function ShoppingSearch() {
 
           {/* ST: Search Bar */}
           <div className="w-fit lg:mt-[1.5625rem] lg:mx-auto">
-            <Search />
+            <SearchBar />
           </div>
           {/* ED: Search Bar */}
